Update providers to restrict access
We have updated our templates to not give this repo access to the shared
services provider, and only use a dns role specific to the account
business unit, so that they cannot amend dns entries of another business
unit.  This change updates the files that have already been created so
that they are in line with the new template.