Add locals required for cert validation
The change ensures that the CNAMES required for
the ACM validation are only created in the corresponding
accounts - the Core Network and Core VPC accounts. Previously,
the CNAMES were duplicated.

This commit also adds the deploy/apply step in the CICD workflow
and updates the apply script to also take variable file options.