Add OPA tests to validate Terraform
Whilst we have protected the providers and backend files with
codeowners, there is still the possibility that someone could attempt to
create a provider or backend in a different file. This opa restricts
that from happening, and starts a framework that we can use to enforce
other restrictions that we want. See
https://github.com/open-policy-agent/conftest/tree/master/examples/hcl2
for more examples of the types of policies we could add.