Add DDoS alarm to Performance Hub
Create an alarm from the DDoSDetected metric
Create a high priority SNS topic
Subscribe to the topic the pagerduty cloudwatch endpoints:
low-priority service for non prod, high-priority service for production.

This will allow us to be notified if any DDoS attacks are detected on
performance hub in the #modernisation-platform-high-priority-alarms
channel.