Add automation to tidy old PRs
Using this action - https://github.com/actions/stale we can automate the tidying up of old PRs.

After 30 days of no activity, the PR will be marked as stale.  If no body comments or removes the stale tag, after a further 10 days the PR will be closed and the branch deleted.  This includes draft PRs.