Add shared modules folder
Following a user request, creating a shared module folder in the
environments repo, this is to enable teams to create modules that
they can use across multiple applications, and for users of other
teams to use, with a view that if the module is successful we can
then move it out to its own repo and add tests, version it etc.

Effectively this will allow users to contribute and write shared
modules when they need them.