Refactor performance hub
This PR is part of the ministryofjustice/modernisation-platform#2521 ticket.

This is to refactor platform files in the account to match the templates. Terraform plan has the same output before refactoring and after refactoring.

Also splitting out the single main tf file into smaller files

Note, the change in the source code tag means the ecs module will be
rebuilt, so these changes should be deployed in a controlled manor to
the environments.