Remove certificate from production
A recent deployment failed because the certificate in production was
expired.  This is an ACM certificate, but because it isn't being used
(production uses an imported Gandi certificate), Amazon doesn't renew
unused certificates, and the expired certificate caused the apply to
fail as the state was unexpected.

This PR adds a count in so that the ACM certificate and the validation
records are not created in production.

Also during this change the AMI version was updated so updating here in
the same PR to avoid 2 PRs which would implement changes and cause
redeployments.  We should see a clean plan with this PR as these changes
have been implemented via terraform locally.