NIT-621 change how nd interface is configured
We are choosing to use a find+replace method to configure the interface. Previous method of editing the XML file using PowerShell does not allow us to write certain characters without escaping them. The application does not conform to XML specication and expects some URI characters like & to remain unescaped

Additionally, some parameters we previously set are not required