Switch to validation record per domain and sub
The main domain record was validation but the subdomain record was not.
This was because both records were being create in the main domain, but
the subdomain resolves to the subdomain hosted zone in
core-vpc-development which doesn't contain the record.

In this change we create 2 validation records one for the main domain,
and one for the subdomain in the core-network-services and
core-vpc-development account respectively.

Following this code as an example - https://github.com/ministryofjustice/modernisation-platform-environments/blob/492699a1bbb682e3abc0e500cfd3befa7cf70c58/terraform/environments/delius-jitbit/lb.tf#L121