Extract module for the API, with tests
Modularising our code means individual modules are not dependent on the
root-level modernisation platform configuration, so we can run it on
other environments, such as the data engineering sandbox or localstack.

The motivation for doing this now is so that we can also run tests on
individual modules using terratest. This should help speed up our
dev-test-cycle. This PR tests that the authorizer allows a valid token
through and denies an invalid token.

We can also run tests on the modules in isolation using terratest. This
should help speed up our dev-test-cycle.