Add explicit depends on
To enable logging to cloudwatch in API gateway, a role needs to be set
on the API gateway account resource.

It seems like terraform gets confused the first time this is run, as
tries to apply default_stage before the role is set on the account.

Hopefully adding this depends_on will prevent this from recurring in the
future.