Use the API to create and delete the data product
The previous approach of deleting the glue table didn't work
becuase it left the files in s3, and in this scenario, athena load
thinks there is something wrong, and we need to reload the data
product.

To prevent this from happening every other run, we can delete &
recreate the data product.