Include CopyObject in our custom cloudtrail
We want to have a record of all writes into our s3 buckets. This allows
us to monitor flow of data through the conformance zones.

Currently we log PutObject, but in the landing-to-raw lambda we copy
files from landing to the data bucket, using a raw/ or fail/ prefix.

I think for these events to be logged, we need to include the CopyObject
event name.