Attempting cross-account iam:PassRole
To create the OpenSearch connector, you need permissions to pass the connector's role to the OpenSearch service.  However, the role is now in a different account and iam:PassRole does not support cross-account passing of roles so we get:
```
User: arn:aws:sts::</>:assumed-role/cloud-platform-irsa-</> is not authorized to perform: iam:PassRole on resource: arn:aws:iam::381491960855:role/probation-search-sagemaker-role because no resource-based policy allows the iam:PassRole action
```

This change (along with a similar change in CP) creates a new "xa-opensearch-to-sagemaker" role in CP with permissions to assume the "probation-search-sagemaker-role" role in MP, which will hopefully permit the iam:PassRole on connector creation.