Force AWS to provision v1 replication configuration
Though this is deprecated, this is necessary to force AWS to use V1 of the replication configuration
V2 does not support DeleteMarkerReplication in cross-region transfer
See https://stackoverflow.com/questions/52854094/status-on-aws-s3-cross-region-replication-delete-operations-behaviour
There's a bit of inconsistency here; [from looking at the destroy statement for the previous replication resource in the terraform plan](https://github.com/ministryofjustice/modernisation-platform-environments/actions/runs/13921880986/job/38956920939#step:9:428), it looks like the previous replication_configuration had a `filter` attribute.
According to the docs, if the `filter` attribute is supplied on creation it should be V2 not V1, but given that the documentation also says that V2 is the version for which the `DeleteMarkerReplication` must be set to Disabled, both of these statements can't be true and result in the behaviour we're seeing.
That is unless the `filter` was added to an old V1 configuration in the background, rather than being supplied on creation and therefore resulting in a V2 instance.
We want a V1 configuration (if nothing else this is clear from the behaviour described), so we shouldn't specify filter on creation, instead we should force V1 by specifying `prefix`