PI-3053 Increase pre-prod initial instance count
We need to test the performance of a bulk data ingestion job, which relies on SageMaker for generating text embeddings. Based on rough estimates, we expect to be able to complete this within a weekend with 4 GPU-accelerated instances. Once confirmed in pre-prod, we will scale back to 1 instance and work on implementing auto-scaling to support future ingestion jobs.

This will require an AWS quota increase from 2 to 5.