Create service role for Acquisitive Crime (#12032)
* Create service role for Acquisitive Crime

Previously, AC were piggy-backing on the role for cmt. Here we create a
new role for them to use so that we can more finely control the
permissions granted to them. We also remove their access to the old
role.

* Use flatten instead of concat

The data object is a set, not a list, so concatenation can't be used.

* Skip role creation in higher environments

AC are not yet testing this in higher environments, and we haven't yet
configured the `trusted_role_arns` to be environment-specific, so we
only create it in lower environments.

* Use shorter environment parsing