Allow dev and test service accounts to assume Rekognition role
Configure a collection of principals to be given permission to
assume the Rekognition role. This group is represented as a
map from a descriptive alias to the principal ARN.

Update the role trust policy document to use a dynamic block
to generate the grant statements from the collection of
principals.

Make the developer_role_suffix variable optional. If specified
a 'Developer' principal is configured within the list of principals
allowed to assume the role.