Redirect requests to use HTTPS in MAAT CloudFront
Incidents and support requests are sometimes raised by caseworkers that
they cannot access MAAT, receiving a HTTP 403 response, which is caused
when they navigate to the app over HTTP and not HTTPS.

CloudFront is the front door and the first point of infra within AWS
when a request is made by a user to MAAT. The `viewer_protocol_policy`
of the CloudFront distribution controls how users can access the origin
and the value is currently set to `https-only`, which means that requests
not receieved over HTTPS will be rejected outright and there is no opportunity
to put in a redirect downstream (say at the load balancer) to redirect to HTTPS,
as CloudFront will not call further down the stack.

This commit changes the value for the `viewer_protocol_policy` of the default
cache behaviour for the distribution from `https-only` to `redirect-to-https`,
which will redirect requests received over HTTP to HTTPS that are not handled
by the specific cache behaviours (currently in place for assets -  images,
stylesheets and JavaScript files). Only an initial request to MAAT sent over
HTTP will need to be redirected to HTTPS, hence only the default behaviour needs
to be modified; the remaining asset-specific behaviours can all keep their
`viewer_protocol_policy` values of `https-only`.

Further reading:

- [CloudFront docs for redirecting HTTP to HTTPS](https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/using-https-viewers-to-cloudfront.html#configure-cloudfront-HTTPS-viewers).